pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrapekit"
version = "1.0.0"
description = "A professional web data extraction toolkit with multi-format export support"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Long Hang", email = "nhutlong20112004@gmail.com" }]
keywords = ["scraping", "web-scraping", "data-extraction", "csv", "json", "excel"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pandas>=2.1",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
    "pydantic>=2.5",
    "loguru>=0.7",
    "tenacity>=8.2",
    "rich>=13.0",
    "aiohttp>=3.9",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "responses>=0.24",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
]
playwright = ["playwright>=1.40"]

[project.scripts]
scrapekit = "scrapekit.cli:main"

[project.urls]
Repository = "https://github.com/longhang/scrapekit"
"Bug Tracker" = "https://github.com/longhang/scrapekit/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["scrapekit*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=scrapekit --cov-report=term-missing -v"
